requirements.txt
tensorflow>=2.15.0
keras>=3.0.0
numpy
opencv-python-headless
pillow
gradio
